requirements.txt
psycopg2-binary
pandas
openpyxl
scikit-learn
schedule
torch
holidays
google-genai
transformers
joblib
numpy==1.26.4
faiss-cpu==1.8.0
python-dotenv
# Additional packages for financial sentiment analysis
accelerate
sentencepiece
protobuf
konlpy
JPype1
